+++ changelog.md
@@ -1,2 +1,1 @@
-fixed a memory leak in codecui 
-added pack info button
+- fixed a config and particle issue